chore: migrate new side bar
